openssl: Update license
The Poky layer now verifies whether the LICENSE value is obsolete or not
(refer to commit 321cf8962ea2 in the poky, kirkston branch). The
"openssl" license, which points to "OpenSSL," is classified as obsolete.
Hence, we replace it with the original name.

According to https://www.openssl.org/source/license.html, the following
licensing information applies:

"For the 3.0 release and later releases derived from that, the Apache
License v2 applies. This also applies to the git 'master' branch. For
any release made before OpenSSL 3.0 (specifically the 1.1.1, 1.1.0,
1.0.2, and all prior releases, including those not currently supported),
the dual OpenSSL and SSLeay license applies."

This is why we are still using the "OpenSSL" license.

Issue GEN-3427